Remove react-native directory
