--- src/react-native/index.ts
@@ -1,1 +1,0 @@
-export { NativeModules } from "react-native";
