--- src/react-native/index.web.ts
@@ -1,1 +1,0 @@
-export { NativeModules } from "./native-modules";
